+++ railties/test/application/loading_test.rb
@@ -111,6 +111,15 @@ class LoadingTest < ActiveSupport::TestCase
     assert ::Rails.application.config.loaded
   end
 
+  test "raises RuntimeError when config/environments/ENV.rb file does not exist" do
+    msg = <<~MSG.squish
+      Rails environment has been set to non_existent_environment but
+      config/environments/non_existent_environment.rb does not exist.
+    MSG
+
+    assert_raise(RuntimeError, match: msg) { boot_app "non_existent_environment" }
+  end
+
   test "descendants loaded after framework initialization are cleaned on each request if reloading is enabled" do
     add_to_config <<-RUBY
       config.enable_reloading = true